requirements.txt
selenium==4.21.0
tqdm==4.66.4
pandas==2.2.2
NRCLex==3.0.0
scikit-learn==1.5.1
streamlit==1.38.0
matplotlib==3.9.2
seaborn==0.13.2
wordcloud==1.9.3
plotly==5.24.0
streamlit-folium==0.22.0
folium==0.17.0

